# backend/api/swagger.yaml
---
basePath: "/v2"
consumes:
- application/json
- application/vnd.api+json
definitions:
  Action:
    example:
      id: id
      page: page
      type: type
      uuid: uuid
      value: "{}"
    properties:
      id:
        description: ID of action
        type: string
      page:
        description: Parent page
        type: string
      type:
        description: Type of action
        type: string
      uuid:
        description: UUID of action
        type: string
      value:
        description: Value of action
        properties: {}
        type: object
    type: object
  Error:
    properties:
      code:
        format: int32
        type: integer
      fields:
        type: string
      message:
        type: string
    type: object
  Exercise:
    example:
      id: id
      lesson: lesson
      name: name
      uuid: uuid
    properties:
      id:
        description: ID of exercise
        type: string
      lesson:
        description: Parent lesson
        type: string
      name:
        description: Name of exercise
        type: string
      uuid:
        description: UUID of exercise
        type: string
    type: object
  Klass:
    example:
      id: id
      name: name
      uuid: uuid
    properties:
      id:
        description: ID of class
        type: string
      name:
        description: Name of class
        type: string
      uuid:
        description: UUID of class
        type: string
    type: object
  Lesson:
    example:
      id: id
      klass: klass
      name: name
      uuid: uuid
    properties:
      id:
        description: ID of lesson
        type: string
      klass:
        description: Parent class
        type: string
      name:
        description: Name of lesson
        type: string
      uuid:
        description: UUID of lesson
        type: string
    type: object
  NAction:
    example:
      page: page
      type: type
      uuid: uuid
      value: "{}"
    properties:
      page:
        description: Parent page
        type: string
      type:
        description: Type of action
        type: string
      uuid:
        description: UUID of action
        type: string
      value:
        description: Value of action
        properties: {}
        type: object
    type: object
  NExercise:
    example:
      lesson: lesson
      name: name
      uuid: uuid
    properties:
      lesson:
        description: Parent lesson
        type: string
      name:
        description: Name of exercise
        type: string
      uuid:
        description: UUID of exercise
        type: string
    type: object
  NKlass:
    example:
      name: name
      uuid: uuid
    properties:
      name:
        description: Name of class
        type: string
      uuid:
        description: UUID of class
        type: string
    type: object
  NLesson:
    example:
      klass: klass
      name: name
      uuid: uuid
    properties:
      klass:
        description: Parent class
        type: string
      name:
        description: Name of lesson
        type: string
      uuid:
        description: UUID of lesson
        type: string
    type: object
  NPage:
    example:
      exercise: exercise
      name: name
      uuid: uuid
    properties:
      exercise:
        description: Parent exercise
        type: string
      name:
        description: Name of page
        type: string
      uuid:
        description: UUID of page
        type: string
    type: object
  NTemplate:
    example:
      collection: collection
      name: name
      type: type
      uuid: uuid
    properties:
      collection:
        description: A list of IDs making up the collection
        items:
          type: string
        type: array
      name:
        description: Name of template
        type: string
      type:
        description: Type of collection
        type: string
      uuid:
        description: UUID of template
        type: string
    type: object
  Page:
    example:
      exercise: exercise
      id: id
      name: name
      uuid: uuid
    properties:
      exercise:
        description: Parent exercise
        type: string
      id:
        description: ID of page
        type: string
      name:
        description: Name of page
        type: string
      uuid:
        description: UUID of page
        type: string
    type: object
  Template:
    example:
      collection: collection
      id: id
      name: name
      type: type
      uuid: uuid
    properties:
      collection:
        description: A list of IDs making up the collection
        items:
          type: string
        type: array
      id:
        description: ID of template
        type: string
      name:
        description: Name of template
        type: string
      type:
        description: Type of collection
        type: string
      uuid:
        description: UUID of template
        type: string
    type: object
info:
  description: WALAT-NEXT
  title: walat
  version: 2.0.0
paths:
  "/actions":
    get:
      description: Returns information about actions at a given page.
      operationId: actionsGET
      parameters:
      - description: Page ID
        in: query
        name: page
        required: false
        type: integer
      responses:
        '200':
          description: An array of actions
          schema:
            items:
              "$ref": "#/definitions/Action"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Action
      x-swagger-router-controller: Action
    post:
      description: Adds a new action to a given page.
      operationId: actionsPOST
      parameters:
      - description: Action with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NAction"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Action
      x-swagger-router-controller: Action
  "/actions/reorder":
    post:
      description: Reorders actions.
      operationId: actionsReorderPOST
      parameters:
      - description: List of ordered action IDs
        in: body
        name: body
        required: true
        schema:
          properties:
            order:
              items:
                type: string
              type: array
          type: object
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Action
      x-swagger-router-controller: Action
  "/actions/{id}":
    delete:
      description: Deletes a action.
      operationId: actionsIdDELETE
      parameters:
      - description: ID of action
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Action
      x-swagger-router-controller: Action
    get:
      description: Returns information about the action of a specified ID.
      operationId: actionsIdGET
      parameters:
      - description: ID of action
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: A action
          schema:
            "$ref": "#/definitions/Action"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Action
      x-swagger-router-controller: Action
    patch:
      description: Edits a action.
      operationId: actionsIdPATCH
      parameters:
      - description: ID of action
        in: path
        name: id
        required: true
        type: string
      - description: Action content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Action"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Action
      x-swagger-router-controller: Action
  "/exercises":
    get:
      description: Returns information about exercises at a given lesson.
      operationId: exercisesGET
      parameters:
      - description: Lesson ID
        in: query
        name: lesson
        required: false
        type: integer
      responses:
        '200':
          description: An array of exercises
          schema:
            items:
              "$ref": "#/definitions/Exercise"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
    post:
      description: Adds a new exercise to a given lesson.
      operationId: exercisesPOST
      parameters:
      - description: Exercise with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NExercise"
      responses:
        '200':
          description: OK
          schema:
            format: int32
            type: integer
        '405':
          description: Invalid input
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
  "/exercises/reorder":
    post:
      description: Reorders exercises.
      operationId: exercisesReorderPOST
      parameters:
      - description: List of ordered exercise IDs
        in: body
        name: body
        required: true
        schema:
          properties:
            order:
              items:
                type: string
              type: array
          type: object
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
  "/exercises/{id}":
    delete:
      description: Deletes an exercise.
      operationId: exercisesIdDELETE
      parameters:
      - description: ID of exercise
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
    get:
      description: Returns information about the exercise of a specified ID.
      operationId: exercisesIdGET
      parameters:
      - description: ID of exercise
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: An exercise
          schema:
            "$ref": "#/definitions/Exercise"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
    patch:
      description: Edits an exercise.
      operationId: exercisesIdPATCH
      parameters:
      - description: ID of exercise
        in: path
        name: id
        required: true
        type: string
      - description: Exercise content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Exercise"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Exercise
      x-swagger-router-controller: Exercise
  "/klasses":
    get:
      description: Returns information about klasses.
      operationId: klassesGET
      responses:
        '200':
          description: An array of klasses
          schema:
            items:
              "$ref": "#/definitions/Klass"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Klass
      x-swagger-router-controller: Klass
    post:
      description: Adds a new klass.
      operationId: klassesPOST
      parameters:
      - description: Klass with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NKlass"
      responses:
        '200':
          description: OK
          schema:
            format: int32
            type: integer
        '405':
          description: Invalid input
      tags:
      - Klass
      x-swagger-router-controller: Klass
  "/klasses/{id}":
    delete:
      description: Deletes a klass.
      operationId: klassesIdDELETE
      parameters:
      - description: ID of klass
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Klass
      x-swagger-router-controller: Klass
    get:
      description: Returns information about the klass of a specified ID.
      operationId: klassesIdGET
      parameters:
      - description: ID of klass
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: A klass
          schema:
            "$ref": "#/definitions/Klass"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Klass
      x-swagger-router-controller: Klass
    patch:
      description: Edits a klass.
      operationId: klassesIdPATCH
      parameters:
      - description: ID of klass
        in: path
        name: id
        required: true
        type: string
      - description: Klass content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Klass"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Klass
      x-swagger-router-controller: Klass
  "/lessons":
    get:
      description: Returns information about lessons.
      operationId: lessonsGET
      parameters:
      - description: Class ID
        in: query
        name: klass
        required: false
        type: integer
      responses:
        '200':
          description: An array of lessons
          schema:
            items:
              "$ref": "#/definitions/Lesson"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
    post:
      description: Adds a new lesson.
      operationId: lessonsPOST
      parameters:
      - description: Lesson with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NLesson"
      responses:
        '200':
          description: OK
          schema:
            format: int32
            type: integer
        '405':
          description: Invalid input
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
  "/lessons/reorder":
    post:
      description: Reorders lessons.
      operationId: lessonsReorderPOST
      parameters:
      - description: List of ordered lesson IDs
        in: body
        name: body
        required: true
        schema:
          properties:
            order:
              items:
                type: string
              type: array
          type: object
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
  "/lessons/{id}":
    delete:
      description: Deletes a lesson.
      operationId: lessonsIdDELETE
      parameters:
      - description: ID of lesson
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
    get:
      description: Returns information about the lesson of a specified ID.
      operationId: lessonsIdGET
      parameters:
      - description: ID of lesson
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: A lesson
          schema:
            "$ref": "#/definitions/Lesson"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
    patch:
      description: Edits a lesson.
      operationId: lessonsIdPATCH
      parameters:
      - description: ID of lesson
        in: path
        name: id
        required: true
        type: string
      - description: Lesson content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Lesson"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Lesson
      x-swagger-router-controller: Lesson
  "/pages":
    get:
      description: Returns information about pages at a given exercise.
      operationId: pagesGET
      parameters:
      - description: Exercise ID
        in: query
        name: exercise
        required: false
        type: integer
      responses:
        '200':
          description: An array of pages
          schema:
            items:
              "$ref": "#/definitions/Page"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Page
      x-swagger-router-controller: Page
    post:
      description: Adds a new page to a given exercise.
      operationId: pagesPOST
      parameters:
      - description: Page with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NPage"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Page
      x-swagger-router-controller: Page
  "/pages/reorder":
    post:
      description: Reorders pages.
      operationId: pagesReorderPOST
      parameters:
      - description: List of ordered page IDs
        in: body
        name: body
        required: true
        schema:
          properties:
            order:
              items:
                type: string
              type: array
          type: object
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Page
      x-swagger-router-controller: Page
  "/pages/{id}":
    delete:
      description: Deletes a page.
      operationId: pagesIdDELETE
      parameters:
      - description: ID of page
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Page
      x-swagger-router-controller: Page
    get:
      description: Returns information about the page of a specified ID.
      operationId: pagesIdGET
      parameters:
      - description: ID of page
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: A page
          schema:
            "$ref": "#/definitions/Page"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Page
      x-swagger-router-controller: Page
    patch:
      description: Edits a page.
      operationId: pagesIdPATCH
      parameters:
      - description: ID of page
        in: path
        name: id
        required: true
        type: string
      - description: Page content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Page"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Page
      x-swagger-router-controller: Page
  "/templates":
    get:
      description: Returns information about templates at a given page.
      operationId: templatesGET
      parameters: []
      responses:
        '200':
          description: An array of templates
          schema:
            items:
              "$ref": "#/definitions/Template"
            type: array
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Template
      x-swagger-router-controller: Template
    post:
      description: Adds a new template to a given page.
      operationId: templatesPOST
      parameters:
      - description: Template with default ID to be added
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/NTemplate"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Template
      x-swagger-router-controller: Template
  "/templates/{id}":
    delete:
      description: Deletes a template.
      operationId: templatesIdDELETE
      parameters:
      - description: ID of template
        in: path
        name: id
        required: true
        type: string
      responses:
        '204':
          description: OK
        '405':
          description: Invalid input
      tags:
      - Template
      x-swagger-router-controller: Template
    get:
      description: Returns information about the template of a specified ID.
      operationId: templatesIdGET
      parameters:
      - description: ID of template
        in: path
        name: id
        required: true
        type: string
      responses:
        '200':
          description: A template
          schema:
            "$ref": "#/definitions/Template"
        default:
          description: Unexpected error
          schema:
            "$ref": "#/definitions/Error"
      tags:
      - Template
      x-swagger-router-controller: Template
    patch:
      description: Edits a template.
      operationId: templatesIdPATCH
      parameters:
      - description: ID of template
        in: path
        name: id
        required: true
        type: string
      - description: Template content to be updated
        in: body
        name: body
        required: true
        schema:
          "$ref": "#/definitions/Template"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '405':
          description: Invalid input
      tags:
      - Template
      x-swagger-router-controller: Template
produces:
- application/json
- application/vnd.api+json
schemes:
- http
swagger: '2.0'